regs_tb.v: add delays before sampling interrupt signal
Because this signal is now implemented with flip-flops on clock posedge,
most of the tests sampled the signal before it actually was changed. As
conditions for signaling interrupt changed on falling clock edge, this
introduces a delay of a half clock cycle, but this is insignificant when
compared to other delays in host's interrupt path, especially for SERIRQ
used on LPC bus.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>