WIP: regs_module.v: switch to combinational logic
This is done as an attempt to get rid of wait state, which seems to be
not supported by Linux SPI TPM driver. It somewhat works for some of
the registers, but not for FIFO, and maybe some other registers that
couldn't be tested without working FIFO.

Some of those changes (e.g. interrupt reg->wire) should be extracted
to main code. It makes timing results better.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>